requirements.txt
python-nmap
bcrypt
pytest